+++ src/Osintgram.py
@@ -4,16 +4,20 @@ import sys
 import urllib
 import os
 import codecs
+from pathlib import Path
 
 import requests
+import ssl
+ssl._create_default_https_context = ssl._create_unverified_context
 
 from geopy.geocoders import Nominatim
 from instagram_private_api import Client as AppClient
-from instagram_private_api import ClientCookieExpiredError, ClientLoginRequiredError, ClientError
+from instagram_private_api import ClientCookieExpiredError, ClientLoginRequiredError, ClientError, ClientThrottledError
 
 from prettytable import PrettyTable
 
 from src import printcolors as pc
+from src import config
 
 
 class Osintgram:
@@ -27,15 +31,26 @@ class Osintgram:
     target = ""
     writeFile = False
     jsonDump = False
+    cli_mode = False
+    output_dir = "output"
 
-    def __init__(self, target, is_file, is_json):
-        u = self.__getUsername__()
-        p = self.__getPassword__()
-        print("\nAttempt to login...")
+
+    def __init__(self, target, is_file, is_json, is_cli, output_dir, clear_cookies):
+        self.output_dir = output_dir or self.output_dir        
+        u = config.getUsername()
+        p = config.getPassword()
+        self.clear_cookies(clear_cookies)
+        self.cli_mode = is_cli
+        if not is_cli:
+          print("\nAttempt to login...")
         self.login(u, p)
         self.setTarget(target)
         self.writeFile = is_file
         self.jsonDump = is_json
+
+    def clear_cookies(self,clear_cookies):
+        if clear_cookies:
+            self.clear_cache()
 
     def setTarget(self, target):
         self.target = target
@@ -44,26 +59,8 @@ class Osintgram:
         self.is_private = user['is_private']
         self.following = self.check_following()
         self.__printTargetBanner__()
-
-    def __getUsername__(self):
-        try:
-            u = open("config/username.conf", "r").read()
-            u = u.replace("\n", "")
-            return u
-        except FileNotFoundError:
-            pc.printout("Error: file \"config/username.conf\" not found!", pc.RED)
-            pc.printout("\n")
-            sys.exit(0)
-
-    def __getPassword__(self):
-        try:
-            p = open("config/pw.conf", "r").read()
-            p = p.replace("\n", "")
-            return p
-        except FileNotFoundError:
-            pc.printout("Error: file \"config/pw.conf\" not found!", pc.RED)
-            pc.printout("\n")
-            sys.exit(0)
+        self.output_dir = self.output_dir + "/" + str(self.target)
+        Path(self.output_dir).mkdir(parents=True, exist_ok=True)
 
     def __get_feed__(self):
         data = []
@@ -166,14 +163,14 @@ class Osintgram:
                 i = i + 1
 
             if self.writeFile:
-                file_name = "output/" + self.target + "_addrs.txt"
+                file_name = self.output_dir + "/" + self.target + "_addrs.txt"
                 file = open(file_name, "w")
                 file.write(str(t))
                 file.close()
 
             if self.jsonDump:
                 json_data['address'] = addrs_list
-                json_file_name = "output/" + self.target + "_addrs.json"
+                json_file_name = self.output_dir + "/" + self.target + "_addrs.json"
                 with open(json_file_name, 'w') as f:
                     json.dump(json_data, f)
 
@@ -216,7 +213,7 @@ class Osintgram:
             file = None
 
             if self.writeFile:
-                file_name = "output/" + self.target + "_captions.txt"
+                file_name = self.output_dir + "/" + self.target + "_captions.txt"
                 file = open(file_name, "w")
 
             for s in captions:
@@ -227,7 +224,7 @@ class Osintgram:
 
             if self.jsonDump:
                 json_data['captions'] = captions
-                json_file_name = "output/" + self.target + "_followings.json"
+                json_file_name = self.output_dir + "/" + self.target + "_followings.json"
                 with open(json_file_name, 'w') as f:
                     json.dump(json_data, f)
 
@@ -255,7 +252,7 @@ class Osintgram:
             posts += 1
 
         if self.writeFile:
-            file_name = "output/" + self.target + "_comments.txt"
+            file_name = self.output_dir + "/" + self.target + "_comments.txt"
             file = open(file_name, "w")
             file.write(str(comments_counter) + " comments in " + str(posts) + " posts\n")
             file.close()
@@ -265,12 +262,54 @@ class Osintgram:
                 'comment_counter': comments_counter,
                 'posts': posts
             }
-            json_file_name = "output/" + self.target + "_comments.json"
+            json_file_name = self.output_dir + "/" + self.target + "_comments.json"
             with open(json_file_name, 'w') as f:
                 json.dump(json_data, f)
 
         pc.printout(str(comments_counter), pc.MAGENTA)
         pc.printout(" comments in " + str(posts) + " posts\n")
+
+    def get_comment_data(self):
+        if self.check_private_profile():
+            return
+
+        pc.printout("Retrieving all comments, this may take a moment...\n")
+        data = self.__get_feed__()
+        
+        _comments = []
+        t = PrettyTable(['POST ID', 'ID', 'Username', 'Comment'])
+        t.align["POST ID"] = "l"
+        t.align["ID"] = "l"
+        t.align["Username"] = "l"
+        t.align["Comment"] = "l"
+
+        for post in data:
+            post_id = post.get('id')
+            comments = self.api.media_n_comments(post_id)
+            for comment in comments:
+                t.add_row([post_id, comment.get('user_id'), comment.get('user').get('username'), comment.get('text')])
+                comment = {
+                        "post_id": post_id,
+                        "user_id":comment.get('user_id'), 
+                        "username": comment.get('user').get('username'),
+                        "comment": comment.get('text')
+                    }
+                _comments.append(comment)
+        
+        print(t)
+        if self.writeFile:
+            file_name = self.output_dir + "/" + self.target + "_comment_data.txt"
+            with open(file_name, 'w') as f:
+                f.write(str(t))
+                f.close()
+        
+        if self.jsonDump:
+            file_name_json = self.output_dir + "/" + self.target + "_comment_data.json"
+            with open(file_name_json, 'w') as f:
+                f.write("{ \"Comments\":[ \n")
+                f.write('\n'.join(json.dumps(comment) for comment in _comments) + ',\n')
+                f.write("]} ")
+
 
     def get_followers(self):
         if self.check_private_profile():
@@ -278,12 +317,26 @@ class Osintgram:
 
         pc.printout("Searching for target followers...\n")
 
+        _followers = []
         followers = []
+
 
         rank_token = AppClient.generate_uuid()
         data = self.api.user_followers(str(self.target_id), rank_token=rank_token)
 
-        for user in data['users']:
+        _followers.extend(data.get('users', []))
+
+        next_max_id = data.get('next_max_id')
+        while next_max_id:
+            sys.stdout.write("\rCatched %i followers" % len(_followers))
+            sys.stdout.flush()
+            results = self.api.user_followers(str(self.target_id), rank_token=rank_token, max_id=next_max_id)
+            _followers.extend(results.get('users', []))
+            next_max_id = results.get('next_max_id')
+
+        print("\n")
+            
+        for user in _followers:
             u = {
                 'id': user['pk'],
                 'username': user['username'],
@@ -311,14 +364,14 @@ class Osintgram:
                 followings_list.append(follow)
 
         if self.writeFile:
-            file_name = "output/" + self.target + "_followers.txt"
+            file_name = self.output_dir + "/" + self.target + "_followers.txt"
             file = open(file_name, "w")
             file.write(str(t))
             file.close()
 
         if self.jsonDump:
             json_data['followers'] = followers
-            json_file_name = "output/" + self.target + "_followers.json"
+            json_file_name = self.output_dir + "/" + self.target + "_followers.json"
             with open(json_file_name, 'w') as f:
                 json.dump(json_data, f)
 
@@ -330,12 +383,25 @@ class Osintgram:
 
         pc.printout("Searching for target followings...\n")
 
+        _followings = []
         followings = []
 
         rank_token = AppClient.generate_uuid()
         data = self.api.user_following(str(self.target_id), rank_token=rank_token)
 
-        for user in data['users']:
+        _followings.extend(data.get('users', []))
+
+        next_max_id = data.get('next_max_id')
+        while next_max_id:
+            sys.stdout.write("\rCatched %i followings" % len(_followings))
+            sys.stdout.flush()
+            results = self.api.user_following(str(self.target_id), rank_token=rank_token, max_id=next_max_id)
+            _followings.extend(results.get('users', []))
+            next_max_id = results.get('next_max_id')
+
+        print("\n")
+
+        for user in _followings:
             u = {
                 'id': user['pk'],
                 'username': user['username'],
@@ -363,14 +429,14 @@ class Osintgram:
                 followings_list.append(follow)
 
         if self.writeFile:
-            file_name = "output/" + self.target + "_followings.txt"
+            file_name = self.output_dir + "/" + self.target + "_followings.txt"
             file = open(file_name, "w")
             file.write(str(t))
             file.close()
 
         if self.jsonDump:
             json_data['followings'] = followings_list
-            json_file_name = "output/" + self.target + "_followings.json"
+            json_file_name = self.output_dir + "/" + self.target + "_followings.json"
             with open(json_file_name, 'w') as f:
                 json.dump(json_data, f)
 
@@ -419,7 +485,7 @@ class Osintgram:
             hashtags_list = []
 
             if self.writeFile:
-                file_name = "output/" + self.target + "_hashtags.txt"
+                file_name = self.output_dir + "/" + self.target + "_hashtags.txt"
                 file = open(file_name, "w")
 
             for k, v in ssort:
@@ -435,7 +501,7 @@ class Osintgram:
 
             if self.jsonDump:
                 json_data['hashtags'] = hashtags_list
-                json_file_name = "output/" + self.target + "_hashtags.json"
+                json_file_name = self.output_dir + "/" + self.target + "_hashtags.json"
                 with open(json_file_name, 'w') as f:
                     json.dump(json_data, f)
         else:
@@ -443,7 +509,7 @@ class Osintgram:
 
     def get_user_info(self):
         try:
-            endpoint = 'users/{user_id!s}/full_detail_info/'.format(**{'user_id': self.target})
+            endpoint = 'users/{user_id!s}/full_detail_info/'.format(**{'user_id': self.target_id})
             content = self.api._call_api(endpoint)
            
             data = content['user_detail']['user']
@@ -494,8 +560,8 @@ class Osintgram:
                     'biography': data['biography'],
                     'edge_followed_by': data['follower_count'],
                     'edge_follow': data['following_count'],
-                    'is_business_account': data['is_business_account'],
-                    'is_verified': data['is_business'],
+                    'is_business_account': data['is_business'],
+                    'is_verified': data['is_verified'],
                     'profile_pic_url_hd': data['hd_profile_pic_url_info']['url']
                 }
                 if 'public_email' in data and data['public_email']:
@@ -511,11 +577,12 @@ class Osintgram:
                 if 'contact_phone_number' in data and data['contact_phone_number']:
                     user['contact_phone_number'] = data['contact_phone_number']
 
-                json_file_name = "output/" + self.target + "_info.json"
+                json_file_name = self.output_dir + "/" + self.target + "_info.json"
                 with open(json_file_name, 'w') as f:
                     json.dump(user, f)
 
         except ClientError as e:
+            print(e)
             pc.printout("Oops... " + str(self.target) + " non exist, please enter a valid username.", pc.RED)
             pc.printout("\n")
             exit(2)
@@ -536,7 +603,7 @@ class Osintgram:
             posts += 1
 
         if self.writeFile:
-            file_name = "output/" + self.target + "_likes.txt"
+            file_name = self.output_dir + "/" + self.target + "_likes.txt"
             file = open(file_name, "w")
             file.write(str(like_counter) + " likes in " + str(like_counter) + " posts\n")
             file.close()
@@ -546,7 +613,7 @@ class Osintgram:
                 'like_counter': like_counter,
                 'posts': like_counter
             }
-            json_file_name = "output/" + self.target + "_likes.json"
+            json_file_name = self.output_dir + "/" + self.target + "_likes.json"
             with open(json_file_name, 'w') as f:
                 json.dump(json_data, f)
 
@@ -581,7 +648,7 @@ class Osintgram:
         if counter > 0:
 
             if self.writeFile:
-                file_name = "output/" + self.target + "_mediatype.txt"
+                file_name = self.output_dir + "/" + self.target + "_mediatype.txt"
                 file = open(file_name, "w")
                 file.write(str(photo_counter) + " photos and " + str(video_counter) + " video posted by target\n")
                 file.close()
@@ -594,7 +661,7 @@ class Osintgram:
                     "photos": photo_counter,
                     "videos": video_counter
                 }
-                json_file_name = "output/" + self.target + "_mediatype.json"
+                json_file_name = self.output_dir + "/" + self.target + "_mediatype.json"
                 with open(json_file_name, 'w') as f:
                     json.dump(json_data, f)
 
@@ -646,14 +713,14 @@ class Osintgram:
             print(t)
 
             if self.writeFile:
-                file_name = "output/" + self.target + "_users_who_commented.txt"
+                file_name = self.output_dir + "/" + self.target + "_users_who_commented.txt"
                 file = open(file_name, "w")
                 file.write(str(t))
                 file.close()
 
             if self.jsonDump:
                 json_data['users_who_commented'] = ssort
-                json_file_name = "output/" + self.target + "_users_who_commented.json"
+                json_file_name = self.output_dir + "/" + self.target + "_users_who_commented.json"
                 with open(json_file_name, 'w') as f:
                     json.dump(json_data, f)
         else:
@@ -714,14 +781,14 @@ class Osintgram:
             print(t)
 
             if self.writeFile:
-                file_name = "output/" + self.target + "_users_who_tagged.txt"
+                file_name = self.output_dir + "/" + self.target + "_users_who_tagged.txt"
                 file = open(file_name, "w")
                 file.write(str(t))
                 file.close()
 
             if self.jsonDump:
                 json_data['users_who_tagged'] = ssort
-                json_file_name = "output/" + self.target + "_users_who_tagged.json"
+                json_file_name = self.output_dir + "/" + self.target + "_users_who_tagged.json"
                 with open(json_file_name, 'w') as f:
                     json.dump(json_data, f)
         else:
@@ -762,14 +829,14 @@ class Osintgram:
                 count += 1
 
             if self.writeFile:
-                file_name = "output/" + self.target + "_photodes.txt"
+                file_name = self.output_dir + "/" + self.target + "_photodes.txt"
                 file = open(file_name, "w")
                 file.write(str(t))
                 file.close()
 
             if self.jsonDump:
                 json_data['descriptions'] = descriptions_list
-                json_file_name = "output/" + self.target + "_descriptions.json"
+                json_file_name = self.output_dir + "/" + self.target + "_descriptions.json"
                 with open(json_file_name, 'w') as f:
                     json.dump(json_data, f)
 
@@ -782,11 +849,15 @@ class Osintgram:
             return
 
         limit = -1
-        pc.printout("How many photos you want to download (default all): ", pc.YELLOW)
-        user_input = input()
+        if self.cli_mode:
+            user_input = ""
+        else:
+            pc.printout("How many photos you want to download (default all): ", pc.YELLOW)
+            user_input = input()
+          
         try:
             if user_input == "":
-                pc.printout("Downloading all photos avaible...\n")
+                pc.printout("Downloading all photos available...\n")
             else:
                 limit = int(user_input)
                 pc.printout("Downloading " + user_input + " photos...\n")
@@ -815,7 +886,7 @@ class Osintgram:
                     counter = counter + 1
                     url = item["image_versions2"]["candidates"][0]["url"]
                     photo_id = item["id"]
-                    end = "output/" + self.target + "_" + photo_id + ".jpg"
+                    end = self.output_dir + "/" + self.target + "_" + photo_id + ".jpg"
                     urllib.request.urlretrieve(url, end)
                     sys.stdout.write("\rDownloaded %i" % counter)
                     sys.stdout.flush()
@@ -827,7 +898,7 @@ class Osintgram:
                         counter = counter + 1
                         url = i["image_versions2"]["candidates"][0]["url"]
                         photo_id = i["id"]
-                        end = "output/" + self.target + "_" + photo_id + ".jpg"
+                        end = self.output_dir + "/" + self.target + "_" + photo_id + ".jpg"
                         urllib.request.urlretrieve(url, end)
                         sys.stdout.write("\rDownloaded %i" % counter)
                         sys.stdout.flush()
@@ -841,12 +912,12 @@ class Osintgram:
         sys.stdout.write(" photos")
         sys.stdout.flush()
 
-        pc.printout("\nWoohoo! We downloaded " + str(counter) + " photos (saved in output/ folder) \n", pc.GREEN)
+        pc.printout("\nWoohoo! We downloaded " + str(counter) + " photos (saved in " + self.output_dir + " folder) \n", pc.GREEN)
 
     def get_user_propic(self):
 
         try:
-            endpoint = 'users/{user_id!s}/full_detail_info/'.format(**{'user_id': self.target})
+            endpoint = 'users/{user_id!s}/full_detail_info/'.format(**{'user_id': self.target_id})
             content = self.api._call_api(endpoint)
 
             data = content['user_detail']['user']
@@ -859,7 +930,7 @@ class Osintgram:
                 URL = data["hd_profile_pic_versions"][items-1]['url']
 
             if URL != "":
-                end = "output/" + self.target + "_propic.jpg"
+                end = self.output_dir + "/" + self.target + "_propic.jpg"
                 urllib.request.urlretrieve(URL, end)
                 pc.printout("Target propic saved in output folder\n", pc.GREEN)
 
@@ -867,8 +938,9 @@ class Osintgram:
                 pc.printout("Sorry! No results found :-(\n", pc.RED)
         
         except ClientError as e:
-            print(e)
-            print("An error occured... exit")
+            error = json.loads(e.error_response)
+            print(error['message'])
+            print(error['error_title'])
             exit(2)
 
     def get_user_stories(self):
@@ -887,12 +959,12 @@ class Osintgram:
                 story_id = i["id"]
                 if i["media_type"] == 1:  # it's a photo
                     url = i['image_versions2']['candidates'][0]['url']
-                    end = "output/" + self.target + "_" + story_id + ".jpg"
+                    end = self.output_dir + "/" + self.target + "_" + story_id + ".jpg"
                     urllib.request.urlretrieve(url, end)
 
                 elif i["media_type"] == 2:  # it's a gif or video
                     url = i['video_versions'][0]['url']
-                    end = "output/" + self.target + "_" + story_id + ".mp4"
+                    end = self.output_dir + "/" + self.target + "_" + story_id + ".mp4"
                     urllib.request.urlretrieve(url, end)
 
         if counter > 0:
@@ -958,14 +1030,14 @@ class Osintgram:
                     tagged_list.append(tag)
 
             if self.writeFile:
-                file_name = "output/" + self.target + "_tagged.txt"
+                file_name = self.output_dir + "/" + self.target + "_tagged.txt"
                 file = open(file_name, "w")
                 file.write(str(t))
                 file.close()
 
             if self.jsonDump:
                 json_data['tagged'] = tagged_list
-                json_file_name = "output/" + self.target + "_tagged.json"
+                json_file_name = self.output_dir + "/" + self.target + "_tagged.json"
                 with open(json_file_name, 'w') as f:
                     json.dump(json_data, f)
 
@@ -975,25 +1047,28 @@ class Osintgram:
 
     def get_user(self, username):
         try:
-            endpoint = 'users/{user_id!s}/full_detail_info/'.format(**{'user_id': self.target})
-            content = self.api._call_api(endpoint)
-
+            content = self.api.username_info(username)
             if self.writeFile:
-                file_name = "output/" + self.target + "_user_id.txt"
+                file_name = self.output_dir + "/" + self.target + "_user_id.txt"
                 file = open(file_name, "w")
-                file.write(str(content['user_detail']['user']['pk']))
+                file.write(str(content['user']['pk']))
                 file.close()
 
             user = dict()
-            user['id'] = content['user_detail']['user']['pk']
-            user['is_private'] = content['user_detail']['user']['is_private']
+            user['id'] = content['user']['pk']
+            user['is_private'] = content['user']['is_private']
 
             return user
         except ClientError as e:
-            pc.printout("Oops... " + str(self.target) + " non exist, please enter a valid username.", pc.RED)
-            pc.printout("\n")
-            exit(2)
-
+            pc.printout('ClientError {0!s} (Code: {1:d}, Response: {2!s})'.format(e.msg, e.code, e.error_response), pc.RED)
+            error = json.loads(e.error_response)
+            if 'message' in error:
+                print(error['message'])
+            if 'error_title' in error:
+                print(error['error_title'])
+            if 'challenge' in error:
+                print("Please follow this link to complete the challenge: " + error['challenge']['url'])    
+            sys.exit(2)
         
 
     def set_write_file(self, flag):
@@ -1025,7 +1100,7 @@ class Osintgram:
             settings_file = "config/settings.json"
             if not os.path.isfile(settings_file):
                 # settings file does not exist
-                print('Unable to find file: {0!s}'.format(settings_file))
+                print(f'Unable to find file: {settings_file!s}')
 
                 # login new
                 self.api = AppClient(auto_patch=True, authenticate=True, username=u, password=p,
@@ -1043,7 +1118,7 @@ class Osintgram:
                     on_login=lambda x: self.onlogin_callback(x, settings_file))
 
         except (ClientCookieExpiredError, ClientLoginRequiredError) as e:
-            print('ClientCookieExpiredError/ClientLoginRequiredError: {0!s}'.format(e))
+            print(f'ClientCookieExpiredError/ClientLoginRequiredError: {e!s}')
 
             # Login expired
             # Do relogin but use default ua, keys and such
@@ -1051,7 +1126,7 @@ class Osintgram:
                                  on_login=lambda x: self.onlogin_callback(x, settings_file))
 
         except ClientError as e:
-            #pc.printout('ClientError {0!s} (Code: {1:d}, Response: {2!s})'.format(e.msg, e.code, e.error_response), pc.RED)
+            pc.printout('ClientError {0!s} (Code: {1:d}, Response: {2!s})'.format(e.msg, e.code, e.error_response), pc.RED)
             error = json.loads(e.error_response)
             pc.printout(error['message'], pc.RED)
             pc.printout(": ", pc.RED)
@@ -1099,28 +1174,78 @@ class Osintgram:
         if self.check_private_profile():
             return
 
-        pc.printout("Searching for emails of target followers... this can take a few minutes\n")
-
         followers = []
+        
+        try:
 
-        rank_token = AppClient.generate_uuid()
-        data = self.api.user_followers(str(self.target_id), rank_token=rank_token)
+            pc.printout("Searching for emails of target followers... this can take a few minutes\n")
 
-        for user in data['users']:
-            u = {
-                'id': user['pk'],
-                'username': user['username'],
-                'full_name': user['full_name']
-            }
-            followers.append(u)
+            rank_token = AppClient.generate_uuid()
+            data = self.api.user_followers(str(self.target_id), rank_token=rank_token)
 
-        results = []
+            for user in data.get('users', []):
+                u = {
+                    'id': user['pk'],
+                    'username': user['username'],
+                    'full_name': user['full_name']
+                }
+                followers.append(u)
 
-        for follow in followers:
-            user = self.api.user_info(str(follow['id']))
-            if 'public_email' in user['user']:
-                follow['email'] = user['user']['public_email']
-                results.append(follow)
+            next_max_id = data.get('next_max_id')
+            while next_max_id:
+                sys.stdout.write("\rCatched %i followers email" % len(followers))
+                sys.stdout.flush()
+                results = self.api.user_followers(str(self.target_id), rank_token=rank_token, max_id=next_max_id)
+                
+                for user in results.get('users', []):
+                    u = {
+                        'id': user['pk'],
+                        'username': user['username'],
+                        'full_name': user['full_name']
+                    }
+                    followers.append(u)
+
+                next_max_id = results.get('next_max_id')
+            
+            print("\n")
+
+            results = []
+            
+            pc.printout("Do you want to get all emails? y/n: ", pc.YELLOW)
+            value = input()
+            
+            if value == str("y") or value == str("yes") or value == str("Yes") or value == str("YES"):
+                value = len(followers)
+            elif value == str(""):
+                print("\n")
+                return
+            elif value == str("n") or value == str("no") or value == str("No") or value == str("NO"):
+                while True:
+                    try:
+                        pc.printout("How many emails do you want to get? ", pc.YELLOW)
+                        new_value = int(input())
+                        value = new_value - 1
+                        break
+                    except ValueError:
+                        pc.printout("Error! Please enter a valid integer!", pc.RED)
+                        print("\n")
+                        return
+            else:
+                pc.printout("Error! Please enter y/n :-)", pc.RED)
+                print("\n")
+                return
+
+            for follow in followers:
+                user = self.api.user_info(str(follow['id']))
+                if 'public_email' in user['user'] and user['user']['public_email']:
+                    follow['email'] = user['user']['public_email']
+                    if len(results) > value:
+                        break
+                    results.append(follow)
+
+        except ClientThrottledError  as e:
+            pc.printout("\nError: Instagram blocked the requests. Please wait a few minutes before you try again.", pc.RED)
+            pc.printout("\n")
 
         if len(results) > 0:
 
@@ -1136,14 +1261,14 @@ class Osintgram:
                 t.add_row([str(node['id']), node['username'], node['full_name'], node['email']])
 
             if self.writeFile:
-                file_name = "output/" + self.target + "_fwersemail.txt"
+                file_name = self.output_dir + "/" + self.target + "_fwersemail.txt"
                 file = open(file_name, "w")
                 file.write(str(t))
                 file.close()
 
             if self.jsonDump:
                 json_data['followers_email'] = results
-                json_file_name = "output/" + self.target + "_fwersemail.json"
+                json_file_name = self.output_dir + "/" + self.target + "_fwersemail.json"
                 with open(json_file_name, 'w') as f:
                     json.dump(json_data, f)
 
@@ -1155,31 +1280,81 @@ class Osintgram:
         if self.check_private_profile():
             return
 
-        pc.printout("Searching for emails of users followed by target... this can take a few minutes\n")
-
         followings = []
 
-        rank_token = AppClient.generate_uuid()
-        data = self.api.user_following(str(self.target_id), rank_token=rank_token)
+        try:
 
-        for user in data['users']:
-            u = {
-                'id': user['pk'],
-                'username': user['username'],
-                'full_name': user['full_name']
-            }
-            followings.append(u)
+            pc.printout("Searching for emails of users followed by target... this can take a few minutes\n")
 
-        results = []
+            rank_token = AppClient.generate_uuid()
+            data = self.api.user_following(str(self.target_id), rank_token=rank_token)
 
-        for follow in followings:
-            user = self.api.user_info(str(follow['id']))
-            if 'public_email' in user['user']:
-                follow['email'] = user['user']['public_email']
-                results.append(follow)
+            for user in data.get('users', []):
+                u = {
+                    'id': user['pk'],
+                    'username': user['username'],
+                    'full_name': user['full_name']
+                }
+                followings.append(u)
+
+            next_max_id = data.get('next_max_id')
+            
+            while next_max_id:
+                results = self.api.user_following(str(self.target_id), rank_token=rank_token, max_id=next_max_id)
+
+                for user in results.get('users', []):
+                    u = {
+                        'id': user['pk'],
+                        'username': user['username'],
+                        'full_name': user['full_name']
+                    }
+                    followings.append(u)
+
+                next_max_id = results.get('next_max_id')
+        
+            results = []
+            
+            pc.printout("Do you want to get all emails? y/n: ", pc.YELLOW)
+            value = input()
+            
+            if value == str("y") or value == str("yes") or value == str("Yes") or value == str("YES"):
+                value = len(followings)
+            elif value == str(""):
+                print("\n")
+                return
+            elif value == str("n") or value == str("no") or value == str("No") or value == str("NO"):
+                while True:
+                    try:
+                        pc.printout("How many emails do you want to get? ", pc.YELLOW)
+                        new_value = int(input())
+                        value = new_value - 1
+                        break
+                    except ValueError:
+                        pc.printout("Error! Please enter a valid integer!", pc.RED)
+                        print("\n")
+                        return
+            else:
+                pc.printout("Error! Please enter y/n :-)", pc.RED)
+                print("\n")
+                return
+
+            for follow in followings:
+                sys.stdout.write("\rCatched %i followings email" % len(results))
+                sys.stdout.flush()
+                user = self.api.user_info(str(follow['id']))
+                if 'public_email' in user['user'] and user['user']['public_email']:
+                    follow['email'] = user['user']['public_email']
+                    if len(results) > value:
+                        break
+                    results.append(follow)
+        
+        except ClientThrottledError as e:
+            pc.printout("\nError: Instagram blocked the requests. Please wait a few minutes before you try again.", pc.RED)
+            pc.printout("\n")
+        
+        print("\n")
 
         if len(results) > 0:
-
             t = PrettyTable(['ID', 'Username', 'Full Name', 'Email'])
             t.align["ID"] = "l"
             t.align["Username"] = "l"
@@ -1192,17 +1367,300 @@ class Osintgram:
                 t.add_row([str(node['id']), node['username'], node['full_name'], node['email']])
 
             if self.writeFile:
-                file_name = "output/" + self.target + "_fwingsemail.txt"
+                file_name = self.output_dir + "/" + self.target + "_fwingsemail.txt"
                 file = open(file_name, "w")
                 file.write(str(t))
                 file.close()
 
             if self.jsonDump:
                 json_data['followings_email'] = results
-                json_file_name = "output/" + self.target + "_fwingsemail.json"
+                json_file_name = self.output_dir + "/" + self.target + "_fwingsemail.json"
                 with open(json_file_name, 'w') as f:
                     json.dump(json_data, f)
 
             print(t)
         else:
             pc.printout("Sorry! No results found :-(\n", pc.RED)
+
+    def get_fwingsnumber(self):
+        if self.check_private_profile():
+            return
+       
+        try:
+
+            pc.printout("Searching for phone numbers of users followed by target... this can take a few minutes\n")
+
+            followings = []
+
+            rank_token = AppClient.generate_uuid()
+            data = self.api.user_following(str(self.target_id), rank_token=rank_token)
+
+            for user in data.get('users', []):
+                u = {
+                    'id': user['pk'],
+                    'username': user['username'],
+                    'full_name': user['full_name']
+                }
+                followings.append(u)
+
+            next_max_id = data.get('next_max_id')
+            
+            while next_max_id:
+                results = self.api.user_following(str(self.target_id), rank_token=rank_token, max_id=next_max_id)
+
+                for user in results.get('users', []):
+                    u = {
+                        'id': user['pk'],
+                        'username': user['username'],
+                        'full_name': user['full_name']
+                    }
+                    followings.append(u)
+
+                next_max_id = results.get('next_max_id')
+       
+            results = []
+        
+            pc.printout("Do you want to get all phone numbers? y/n: ", pc.YELLOW)
+            value = input()
+            
+            if value == str("y") or value == str("yes") or value == str("Yes") or value == str("YES"):
+                value = len(followings)
+            elif value == str(""):
+                print("\n")
+                return
+            elif value == str("n") or value == str("no") or value == str("No") or value == str("NO"):
+                while True:
+                    try:
+                        pc.printout("How many phone numbers do you want to get? ", pc.YELLOW)
+                        new_value = int(input())
+                        value = new_value - 1
+                        break
+                    except ValueError:
+                        pc.printout("Error! Please enter a valid integer!", pc.RED)
+                        print("\n")
+                        return
+            else:
+                pc.printout("Error! Please enter y/n :-)", pc.RED)
+                print("\n")
+                return
+
+            for follow in followings:
+                sys.stdout.write("\rCatched %i followings phone numbers" % len(results))
+                sys.stdout.flush()
+                user = self.api.user_info(str(follow['id']))
+                if 'contact_phone_number' in user['user'] and user['user']['contact_phone_number']:
+                    follow['contact_phone_number'] = user['user']['contact_phone_number']
+                    if len(results) > value:
+                        break
+                    results.append(follow)
+
+        except ClientThrottledError as e:
+            pc.printout("\nError: Instagram blocked the requests. Please wait a few minutes before you try again.", pc.RED)
+            pc.printout("\n")
+        
+        print("\n")
+
+        if len(results) > 0:
+            t = PrettyTable(['ID', 'Username', 'Full Name', 'Phone'])
+            t.align["ID"] = "l"
+            t.align["Username"] = "l"
+            t.align["Full Name"] = "l"
+            t.align["Phone number"] = "l"
+
+            json_data = {}
+
+            for node in results:
+                t.add_row([str(node['id']), node['username'], node['full_name'], node['contact_phone_number']])
+
+            if self.writeFile:
+                file_name = self.output_dir + "/" + self.target + "_fwingsnumber.txt"
+                file = open(file_name, "w")
+                file.write(str(t))
+                file.close()
+
+            if self.jsonDump:
+                json_data['followings_phone_numbers'] = results
+                json_file_name = self.output_dir + "/" + self.target + "_fwingsnumber.json"
+                with open(json_file_name, 'w') as f:
+                    json.dump(json_data, f)
+
+            print(t)
+        else:
+            pc.printout("Sorry! No results found :-(\n", pc.RED)
+
+    def get_fwersnumber(self):
+        if self.check_private_profile():
+            return
+
+        followings = []
+
+        try:
+
+            pc.printout("Searching for phone numbers of users followers... this can take a few minutes\n")
+
+
+            rank_token = AppClient.generate_uuid()
+            data = self.api.user_following(str(self.target_id), rank_token=rank_token)
+
+            for user in data.get('users', []):
+                u = {
+                    'id': user['pk'],
+                    'username': user['username'],
+                    'full_name': user['full_name']
+                }
+                followings.append(u)
+
+            next_max_id = data.get('next_max_id')
+            
+            while next_max_id:
+                results = self.api.user_following(str(self.target_id), rank_token=rank_token, max_id=next_max_id)
+
+                for user in results.get('users', []):
+                    u = {
+                        'id': user['pk'],
+                        'username': user['username'],
+                        'full_name': user['full_name']
+                    }
+                    followings.append(u)
+
+                next_max_id = results.get('next_max_id')
+        
+            results = []
+            
+            pc.printout("Do you want to get all phone numbers? y/n: ", pc.YELLOW)
+            value = input()
+            
+            if value == str("y") or value == str("yes") or value == str("Yes") or value == str("YES"):
+                value = len(followings)
+            elif value == str(""):
+                print("\n")
+                return
+            elif value == str("n") or value == str("no") or value == str("No") or value == str("NO"):
+                while True:
+                    try:
+                        pc.printout("How many phone numbers do you want to get? ", pc.YELLOW)
+                        new_value = int(input())
+                        value = new_value - 1
+                        break
+                    except ValueError:
+                        pc.printout("Error! Please enter a valid integer!", pc.RED)
+                        print("\n")
+                        return
+            else:
+                pc.printout("Error! Please enter y/n :-)", pc.RED)
+                print("\n")
+                return
+
+            for follow in followings:
+                sys.stdout.write("\rCatched %i followers phone numbers" % len(results))
+                sys.stdout.flush()
+                user = self.api.user_info(str(follow['id']))
+                if 'contact_phone_number' in user['user'] and user['user']['contact_phone_number']:
+                    follow['contact_phone_number'] = user['user']['contact_phone_number']
+                    if len(results) > value:
+                        break
+                    results.append(follow)
+
+        except ClientThrottledError as e:
+            pc.printout("\nError: Instagram blocked the requests. Please wait a few minutes before you try again.", pc.RED)
+            pc.printout("\n")
+
+        print("\n")
+
+        if len(results) > 0:
+            t = PrettyTable(['ID', 'Username', 'Full Name', 'Phone'])
+            t.align["ID"] = "l"
+            t.align["Username"] = "l"
+            t.align["Full Name"] = "l"
+            t.align["Phone number"] = "l"
+
+            json_data = {}
+
+            for node in results:
+                t.add_row([str(node['id']), node['username'], node['full_name'], node['contact_phone_number']])
+
+            if self.writeFile:
+                file_name = self.output_dir + "/" + self.target + "_fwersnumber.txt"
+                file = open(file_name, "w")
+                file.write(str(t))
+                file.close()
+
+            if self.jsonDump:
+                json_data['followings_phone_numbers'] = results
+                json_file_name = self.output_dir + "/" + self.target + "_fwerssnumber.json"
+                with open(json_file_name, 'w') as f:
+                    json.dump(json_data, f)
+
+            print(t)
+        else:
+            pc.printout("Sorry! No results found :-(\n", pc.RED)
+
+    def get_comments(self):
+        if self.check_private_profile():
+            return
+
+        pc.printout("Searching for users who commented...\n")
+
+        data = self.__get_feed__()
+        users = []
+
+        for post in data:
+            comments = self.__get_comments__(post['id'])
+            for comment in comments:
+                print(comment['text'])
+                
+                # if not any(u['id'] == comment['user']['pk'] for u in users):
+                #     user = {
+                #         'id': comment['user']['pk'],
+                #         'username': comment['user']['username'],
+                #         'full_name': comment['user']['full_name'],
+                #         'counter': 1
+                #     }
+                #     users.append(user)
+                # else:
+                #     for user in users:
+                #         if user['id'] == comment['user']['pk']:
+                #             user['counter'] += 1
+                #             break
+
+        if len(users) > 0:
+            ssort = sorted(users, key=lambda value: value['counter'], reverse=True)
+
+            json_data = {}
+
+            t = PrettyTable()
+
+            t.field_names = ['Comments', 'ID', 'Username', 'Full Name']
+            t.align["Comments"] = "l"
+            t.align["ID"] = "l"
+            t.align["Username"] = "l"
+            t.align["Full Name"] = "l"
+
+            for u in ssort:
+                t.add_row([str(u['counter']), u['id'], u['username'], u['full_name']])
+
+            print(t)
+
+            if self.writeFile:
+                file_name = self.output_dir + "/" + self.target + "_users_who_commented.txt"
+                file = open(file_name, "w")
+                file.write(str(t))
+                file.close()
+
+            if self.jsonDump:
+                json_data['users_who_commented'] = ssort
+                json_file_name = self.output_dir + "/" + self.target + "_users_who_commented.json"
+                with open(json_file_name, 'w') as f:
+                    json.dump(json_data, f)
+        else:
+            pc.printout("Sorry! No results found :-(\n", pc.RED)
+
+    def clear_cache(self):
+        try:
+            f = open("config/settings.json",'w')
+            f.write("{}")
+            pc.printout("Cache Cleared.\n",pc.GREEN)
+        except FileNotFoundError:
+            pc.printout("Settings.json don't exist.\n",pc.RED)
+        finally:
+            f.close()