+++ doc/CHANGELOG.md
@@ -1,5 +1,47 @@
 # Changelog
 
+## [1.3](https://github.com/Datalux/Osintgram/releases/tag/1.3)
+**Enhancements**
+- Artwork refactoring (#149) 
+- Added command line mode (#155)
+- Added output limiter (#201)
+
+**Bug fixes**
+- Losing collected data (#156)
+- JSON user info (#202)
+- Issue #198 (#200)
+- Issue #204 (12e730e)
+
+
+## [1.2](https://github.com/Datalux/Osintgram/releases/tag/1.2)
+**Enhancements**
+- Added virtual environment (#126)
+- Removed some typos (#129, #118)
+- Added new configuration (#125)
+- Added new `commentdata` command (#131) 
+- Added Docker support (#141) 
+
+
+**Bug fixes**
+- Fix bug  #138 (fc2a6be)
+- SSL certificate error (#136) 
+
+
+## [1.1](https://github.com/Datalux/Osintgram/releases/tag/1.1)
+**Enhancements**
+- Improved command parser (#86)
+- Improved errors handling (8bd1abc)
+- Add new line when input command is empty (f5211eb)
+- Added new commands to catch phone number of users (#111)
+- Added support for Windows (#100)
+
+
+**Bug fixes**
+- Fix commands output limit bug (#87)
+- Fix setting target with "." in username (9082990)
+- Readline installing error (#94 )
+
+
 ## [1.0.1](https://github.com/Datalux/Osintgram/releases/tag/1.0.1)
 **Bug fixes**
 - Set itself as target by param
@@ -8,10 +50,10 @@
 **Enhancements**
 - Set itself as target (#53)
 -  Get others info from user (`info` command):
-  - Whats'App number (if avaible)
-  - City Name (if avaible)
-  - Address Street (if avaible)
-  - Contact phone number (if avaible)
+  - Whats'App number (if available)
+  - City Name (if available)
+  - Address Street (if available)
+  - Contact phone number (if available)
 
 **Bug fixes**
 - Fix login issue (#79, #80, #81)
@@ -86,7 +128,7 @@
 
 **Bug fixes**
 
-- added a check if target has a private profile to avoid tool crash (#10)
+- added a check if the target has a private profile to avoid tool crash (#10)
 - fixed `tagged` bug (#5)
 
 ## [0.3](https://github.com/Datalux/Osintgram/releases/tag/0.3)